refactor(aferição): altera tipo de parâmetro de bigint para integer
PL-3570

+++ sql/prolog_queries/frota/afericao/func_afericao_get_afericoes_avulsas_paginada.sql
@@ -1,8 +1,8 @@
 create or replace function func_afericao_get_afericoes_avulsas_paginada(f_cod_unidades bigint[],
                                                                         f_data_inicial date,
                                                                         f_data_final date,
-                                                                        f_limit bigint,
-                                                                        f_offset bigint,
+                                                                        f_limit integer,
+                                                                        f_offset integer,
                                                                         f_incluir_medidas boolean default false)
     returns table
             (
