+++ sql/prolog_queries/frota/afericao/func_afericao_get_afericoes_placas_paginada.sql
@@ -3,8 +3,8 @@ create or replace function func_afericao_get_afericoes_placas_paginada(f_cod_uni
                                                                        f_cod_veiculo bigint,
                                                                        f_data_inicial date,
                                                                        f_data_final date,
-                                                                       f_limit bigint,
-                                                                       f_offset bigint,
+                                                                       f_limit integer,
+                                                                       f_offset integer,
                                                                        f_incluir_medidas boolean default false)
     returns table
             (
